requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
ffmpeg-python==0.2.0
Flask==3.1.0
future==1.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
numpy==2.2.3
openai==1.62.0
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyparsing==3.2.1
requests==2.32.3
rsa==4.9
scipy==1.15.2
sniffio==1.3.1
sounddevice==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
